pyproject.toml
[project]
name = "whale-sounds"
version = "0.1.0"
description = "Simple ambient synth for the Gourd"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gpiozero>=2.0.1", # TODO: remove if not reading GPIO directly into RasPi
    "lgpio>=0.2.2.0", # TODO: remove if not reading GPIO directly into RasPi
    "pigpio>=1.78", # TODO: remove if not reading GPIO directly into RasPi
    "pyo>=1.0.5",
    "rpi-gpio>=0.7.1", # TODO: remove if not reading GPIO directly into RasPi
    "spidev>=3.7", # TODO: remove if not reading GPIO directly into RasPi
]
